Maint: Don't load rubygems in application code
Previously, loading the `Facter::Util::Config` module had the side
effect of loading rubygems, which leads to unpredictable behavior when
facter is used as a library (as opposed to an application via
bin/facter).

This commit removes the require 'rubygems' call so that an external
application can require 'facter' without the ill side-effect.