(#21518) Use WMI to collect per-interface facts
Previously, we used netsh.exe to collect the interfaces fact, and the
derived per-interface facts on Windows. However, netsh.exe is not
available on plain 2003 servers, and the facts relied on netsh.exe
outputting en_US text.

This commit refactors the method to use WMI instead, which does not
have these issues.