(FACT-3488) Fix potential NoMethodError
Commit 7bc38ccb58e6abf66d6141a3b2e9f60eb0077d71 combined the call to
`FileHelper.safe_read(`/etc/machine-id`, nil)` and `strip`. However, if we ever
fail to read `/etc/machine-id`, then we would attempt to call `strip` on nil.

We could use safe-nav operator, however, that would result in the hypervisor
containing:

    { systemd_nspawn: { 'id' => nil } }

This commit restores the original behavior of only including hypervisor
information for `systemd-nspawn` if the machine_id is not nil and protects
against the `nil` case.